rewrite parser
to be able to parse more complicated namelist files.

Valid logical values in other formats than .false. and F are recognised now.
The parser relies in parts on the ast.literal_eval() function.

The result namelist.groups is not case-insensitive, i.e. the keys 'mode' and 'MODE'
refer to the same item.

Treatment of comments is much improved.